Config option to stop menu extending outside a parent element
If `positionMenuWithin` is specified, the width of the menu is
artificially constrained, such that it does not extend outside the
bounds of the `positionMenuWithin` element when it is shown.

This was the first test that really required CSS styles (for the width
and position calculations) so it necessitated a tidying up of the CSS
files in /demo.