Remove `for` attribute from labels when creating a multiSelect
It was noted that it’s bad practice to leave a `<label>` pointing at a
hidden form element. So, if the `<select>` we’re converting is
referenced in the `for` attribute of any `<label>`s in the document,
we remove those `for` attribute, so the label isn’t left pointing at a
hidden form element.

Fixes #28.