Changed markup when using optgroup element
Instead of using the pseudo element to display the title of a optgroup,
now they will be wrapped on a fieldset with a legend. The legend will
have the text from the label attribute from the optgroup.

It's also possible to change the class names for the fielset and legend,
by using these settings: menuFieldsetHTML and menuFieldsetLegendHTML.

Adapt up/down code to work with fieldsets.